Shoot projectiles in the right direction
Changed the shooting orientation depending on how the player is faced. For example if the player is moving left or standing left-faced it will be able to shoot projectiles in his left side. In order to do this I changed the player animations by modifing the firepoint's position(if the player is oriented to the left the firepoint will be on his left side).