BossFight Demo
Added a skeleton spriete, created a folder with characters in the Art folder and made a script for the skeleton. I created a scene with a one door room prefab and a player and I added 2 skeletons in the scene. I also made animations for every move of the skeleton. In order to make the skeleton move in the scene randomly in the scene I added some movepoints. I also added the skeleton tag.